small fixes to quality checks of variants

+++ src/haplotype/haplotype_caller_genotyping_engine.rs
@@ -301,19 +301,23 @@ impl HaplotypeCallerGenotypingEngine {
                             );
                             debug!("Annotated call {:?}", &annotated_call);
 
-                            return_calls.push(annotated_call);
-                            call.alleles
-                                .into_iter()
-                                .enumerate()
-                                .map(|(idx, a)| allele_mapper.remove(&idx))
-                                .for_each(|a| {
-                                    match a {
-                                        None => {
-                                            // do nothing
+                            if annotated_call.get_genotypes().genotypes().iter().any(|g| g.dp - g.ad[0]).sum::<i64>() >= 2 {
+                                // at least two supporting reads
+                                return_calls.push(annotated_call);
+                                call.alleles
+                                    .into_iter()
+                                    .enumerate()
+                                    .map(|(idx, a)| allele_mapper.remove(&idx))
+                                    .for_each(|a| {
+                                        match a {
+                                            None => {
+                                                // do nothing
+                                            }
+                                            Some(a) => called_haplotypes.extend(a.into_iter().cloned()),
                                         }
-                                        Some(a) => called_haplotypes.extend(a.into_iter().cloned()),
-                                    }
-                                });
+                                    });
+                            }
+
                         }
                     }
                 }
